feat: separate flashblocks extension crate
